torrentwatch 0.6-6   * modified add_torrent in torrentwatch.php to use 'Save Torrents' option   * modified display_global_settings to display new option   * modified parse_options() in tw-iface.php to handle new option   * add function setup_default_config() called just before cron installer   * modified tw-iface.php to send different css sheet if environment var   REMOTE_ADDR=127.0.0.1   * Duplicated tw-iface.css to tw-iface.local.css   * Adjust torrentwatch-installer.cgi to link tw-iface.local.css   * Renamed 'Torrent Dir' to 'Watch Dir' to avoid confusion   * Special case in setup_default_config() to move the setting

+++ trunk/torrentwatch.php
@@ -45,7 +45,10 @@ function add_torrent($filename, $dest) {
 		_debug("Starting: $filename\n");
 	else
 		_debug("Failed Starting: $filename\n");
-	unlink($filename);
+	if($config_values['Settings']['Save Torrents'])
+		rename($filename, "$dest/".basename($filename));
+	else
+		unlink($filename);
 }
 
 function check_for_torrents($directory, $dest) {
@@ -69,12 +72,12 @@ function check_for_torrents($directory, $dest) {
 
 read_config_file();
 parse_options();
-if(!isset($config_values['Settings']['Torrent Dir']) or
+if(!isset($config_values['Settings']['Watch Dir']) or
 		!isset($config_values['Settings']['Download Dir'])) {
 	_debug("torrentwatch.php: Bad Config\n\n", 0);
 	exit(1);
 }
-check_for_torrents($config_values['Settings']['Torrent Dir'], $config_values['Settings']['Download Dir']);
+check_for_torrents($config_values['Settings']['Watch Dir'], $config_values['Settings']['Download Dir']);
 if(!$hit)
 	_debug("No New Torrents to add\n", 0);
 exit(0);
